refactor: controllers, films, users по замечаниям ревью 1 (ТЗ-9)
