+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -4,6 +4,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
+import ru.yandex.practicum.filmorate.exception.DoesNotExistException;
 import ru.yandex.practicum.filmorate.model.User;
 
 import javax.validation.Valid;
@@ -14,18 +15,18 @@ import java.util.Map;
 
 @Slf4j
 @RestController
-@RequestMapping
+@RequestMapping("/users")
 public class UserController {
     private int id;
     private final Map<Integer, User> users = new HashMap<>();
 
-    @GetMapping("/users")
+    @GetMapping()
     public List<User> getAllUsers() {
         log.info("Запрос данных всех пользователей");
         return new ArrayList<>(users.values());
     }
 
-    @PostMapping(value = "/users")
+    @PostMapping()
     public User createUser(@Valid @RequestBody User user) {
         if (user.getName() == null || user.getName().isBlank()) {
             user.setName(user.getLogin());
@@ -37,7 +38,7 @@ public class UserController {
         return user;
     }
 
-    @PutMapping(value = "/users")
+    @PutMapping()
     public User updateUser(@RequestBody User user) {
         try {
             if (user.getName() == null || user.getName().isBlank()) {
