+++ src/main/java/ru/yandex/practicum/filmorate/exception/DoesNotExistException.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.controller;
+package ru.yandex.practicum.filmorate.exception;
 
 //класс DoesNotExistException для обработки апдейта несуществующего объекта
 public class DoesNotExistException extends RuntimeException {
