+++ src/main/java/ru/yandex/practicum/filmorate/exception/ResponseError.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.controller;
+package ru.yandex.practicum.filmorate.exception;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Getter;
