+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -2,7 +2,7 @@ package ru.yandex.practicum.filmorate.model;
 
 import lombok.Builder;
 import lombok.Data;
-import ru.yandex.practicum.filmorate.controller.ReleaseValidation;
+import ru.yandex.practicum.filmorate.validators.ReleaseValidation;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Positive;
