+++ src/main/java/ru/yandex/practicum/filmorate/validators/ReleaseValidation.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.controller;
+package ru.yandex.practicum.filmorate.validators;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
