+++ src/main/java/ru/yandex/practicum/filmorate/validators/ReleaseValidator.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.controller;
+package ru.yandex.practicum.filmorate.validators;
 
 import ru.yandex.practicum.filmorate.model.Film;
 
@@ -7,7 +7,9 @@ import javax.validation.ConstraintValidatorContext;
 import java.time.LocalDate;
 
 //класс обработки аннотации @ReleaseValidation
-public class ReleaseValidator implements ConstraintValidator<ReleaseValidation, LocalDate> {
+public class
+
+ReleaseValidator implements ConstraintValidator<ReleaseValidation, LocalDate> {
 
     @Override
     public boolean isValid(LocalDate date, ConstraintValidatorContext context) {
